refactor(capabilities): symmetric v8/v9 simulators over a shared core
Behaviour-preserving restructure of the simulation module introduced earlier
in this branch. All 128 unit tests pass unmodified.

Name both versions, instead of one. The ledger-v9 simulator was implicitly
the unnamed one at the top of `simulation/` while only its ledger-v8 twin was
labelled, which made "which is which" a question you had to answer by reading
imports. Both now sit in `v8/` and `v9/`, reachable as the `V8` and `V9`
namespaces, with the v9 names still exported unqualified so existing code —
and the published `/simulation` subpath — is unaffected.

Share what does not depend on a ledger. `core/` holds the parts of the
simulator that never name a ledger type, written once rather than duplicated
per version:

- VersionTimeline.ts — the protocol-version timeline. Stated over the
  smallest structure it needs, so it applies to any simulator state without
  knowing which ledger the rest of that state belongs to.
- blocks.ts — block hashing and block context. `BlockContext` is a plain
  record defined identically by both ledgers, so it is stated once here and a
  value of it is accepted by either.
- strictness.ts — the strictness configuration and its presets, which are
  plain data. Turning one into the ledger's `WellFormedStrictness` object
  stays per-version, because that constructs a ledger object.

Each version's `SimulatorState` re-exports the shared names, so both barrels
still present one complete surface and no caller learns that core exists.

What deliberately stays duplicated is what actually differs: constructing and
applying ledger objects — genesis minting, night rewards, transaction
application, ledger state. Sharing those needs the simulator to become
generic over roughly seven ledger types, whose cost in signature noise is not
worth paying yet; the twins remain identical apart from their import line, so
the v8 side is still regenerated from the v9 side rather than hand-edited.
